backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
anthropic==0.28.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
python-dotenv==1.0.1
